pyproject.toml
[project]
name = "injection-detector-training"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.6.0,<3.0.0",
    "transformers>=4.35.0,<5.0.0",
    "datasets>=2.14.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "accelerate>=0.24.0",
    "safetensors>=0.4.0",
    "pyyaml>=6.0",
    "pandas>=2.0.3",
    "matplotlib>=3.7.5",
    "seaborn>=0.13.2",
    "plotly>=6.3.0",
    "google-cloud-storage>=3.4.0",
    "huggingface-hub>=0.34.4",
    "notebook>=7.3.3",
    "gcsfs>=2024.9.0.post1",
    "datasketch>=1.6.5",
    "tqdm>=4.67.1",
    "pyarrow>=17.0.0",
]

[dependency-groups]
train = ["torch>=2.0", "transformers>=4.40", "datasets>=2.18", "scikit-learn>=1.4", "accelerate>=0.28", "safetensors>=0.4", "pyyaml>=6.0"]
cloud = ["google-cloud-aiplatform>=1.50", "google-cloud-storage>=2.16", "pyyaml>=6.0"]
export = ["onnx>=1.16", "onnxruntime>=1.18"]
dev = ["pytest>=8.0", "ruff>=0.4"]

[tool.uv]
dev-dependencies = []
